完善健全和参数校验 对图片相关接口做了缓存优化

+++ biz/router/action/action.go
@@ -11,14 +11,16 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+	"regexp"
 	"strings"
 )
 
 // UploadImg 上传图片
 func UploadImg(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	// 从请求中获取上传的文件
 	file, err := c.FormFile("file")
 	extension := filepath.Ext(file.Filename)
@@ -26,32 +28,42 @@ func UploadImg(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"message": "文件上传失败"})
 		return
 	}
-
 	// 创建目标目录
 	folder := c.PostForm("group") // 指定目录分层
 	folderPath := filepath.Join(setting.StorageDir, folder)
 
+	// 删除图片缓存
+	go middleware.DeleteCache(folder)
+
 	// 获取文件名称
 	name := c.PostForm("name")
 
+	// 定义正则表达式模式
+	pattern := "[\\p{P}\\s]+" // 匹配标点符号和空格
+
+	// 编译正则表达式
+	reg := regexp.MustCompile(pattern)
+
+	// 检查字符串是否含有标点符号或空格
+	if reg.MatchString(name) {
+		c.JSON(http.StatusBadRequest, gin.H{"message": "不合法的文件名"})
+	}
 	// 不传name则使用随机串
 	if name == "" {
 		name, _ = middleware.GenerateRandomString(8)
 	}
-
 	// 如果有后缀则去掉后缀
 	if strings.ContainsAny(name, ".") {
 		name = middleware.GetFileNameWithoutExtension(name)
 	}
-
 	// 拼接图片全名
 	fullName := name + extension
 
+	// 递归创建目录
 	if err := os.MkdirAll(folderPath, os.ModePerm); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"message": "目录创建失败"})
 		return
 	}
-
 	// 创建目标文件路径
 	dstPath := filepath.Join(folderPath, fullName)
 
@@ -80,24 +92,29 @@ func UploadImg(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"message": "文件上传成功"})
 }
 
-// ShowImg 下载图片
+// ShowImg 获取图片
 func ShowImg(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
-	// 获取参数凭借路径
+	if middleware.CheckRole(c) == false {
+		return
+	}
+	// 获取参数并尝试从缓存获取结果
 	group := c.Query("group")
+	filePathsCache := middleware.GetCache(group)
+	if filePathsCache != nil {
+		c.JSON(http.StatusOK, gin.H{"filePaths": filePathsCache})
+		return
+	}
+	// 拼接文件夹路径
 	folderPath := filepath.Join(setting.StorageDir, group)
 
 	// 遍历文件夹下的所有文件
 	filePaths := make([]result.Image, 0)
-
 	err := filepath.Walk(folderPath, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			log.Printf("Failed to walk directory: %v\n", err)
 			return nil
 		}
-
 		// 如果是文件，将文件路径添加到切片中
 		if !info.IsDir() {
 			path := setting.Prefix + strings.Replace(path, setting.StorageDir, setting.Proxy, 1)
@@ -110,27 +127,32 @@ func ShowImg(c *gin.Context) {
 		}
 		return nil
 	})
-
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"message": "文件遍历失败"})
 		return
 	}
-
+	// 添加缓存
+	go middleware.AddCache(group, filePaths)
 	c.JSON(http.StatusOK, gin.H{"filePaths": filePaths})
 }
 
 // DeleteImg 删除图片
 func DeleteImg(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	// 获取文件或目录路径
 	path := c.Query("path")
+
+	// 删除图片所在缓存区的缓存
+	index := strings.LastIndex(path, "/")
+	go middleware.DeleteCache(path[:index+1])
 
 	// 拼接完整路径
 	filePath := filepath.Join(setting.StorageDir, path)
 
-	// 删除文件
+	// 删除单个文件
 	err := os.Remove(filePath)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"message": "删除文件失败"})
@@ -142,17 +164,16 @@ func DeleteImg(c *gin.Context) {
 // AddDir 创建目录
 func AddDir(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	var req param.Group
-
-	// 创建目标目录
 	if err := c.ShouldBindJSON(&req); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
 		return
 	}
+	// 递归创建目标目录
 	folderPath := filepath.Join(setting.StorageDir, req.Name)
-
 	if err := os.MkdirAll(folderPath, os.ModePerm); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"message": "目录创建失败"})
 		return
@@ -163,8 +184,9 @@ func AddDir(c *gin.Context) {
 // DeleteDir 删除目录
 func DeleteDir(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	// 获取文件或目录路径
 	path := c.Query("path")
 
@@ -183,8 +205,9 @@ func DeleteDir(c *gin.Context) {
 // ShowRoot 获取目录
 func ShowRoot(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	// 指定目录路径
 	dirPath := setting.StorageDir
 
@@ -198,7 +221,6 @@ func ShowRoot(c *gin.Context) {
 			log.Printf("Encountered error: %v\n", err)
 			return nil
 		}
-
 		if info.IsDir() && path != dirPath {
 			// 获取相对路径
 			relPath, err := filepath.Rel(dirPath, path)
@@ -218,7 +240,6 @@ func ShowRoot(c *gin.Context) {
 		}
 		return nil
 	})
-
 	if err != nil {
 		// 处理遍历目录的错误
 		log.Printf("Failed to walk directory: %v\n", err)
@@ -231,8 +252,9 @@ func ShowRoot(c *gin.Context) {
 // ShowDir 获取目录
 func ShowDir(c *gin.Context) {
 	// 鉴权
-	middleware.CheckRole(c)
-
+	if middleware.CheckRole(c) == false {
+		return
+	}
 	// 指定目录路径
 	path := c.Query("group")
 
@@ -260,7 +282,6 @@ func ShowDir(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, gin.H{"message": "读取目录失败"})
 		return
 	}
-
 	// 遍历目录中的文件和子目录
 	for _, info := range fileInfo {
 		if info.IsDir() {
